backend/requirements.txt
uvicorn
fastapi
python-dotenv
langchain
langchain_openai
python_dotenv